feat: update to Google Maps iOS SDK 6.0.0
