+++ Cartfile
@@ -1,2 +1,2 @@
-#binary "https://dl.google.com/geosdk/GoogleMaps.json" == 5.1.0
-binary "https://dl.google.com/geosdk/GooglePlaces.json" == 5.1.0
+binary "https://dl.google.com/geosdk/GoogleMaps.json" == 6.0.0
+binary "https://dl.google.com/geosdk/GooglePlaces.json" == 6.0.0
