+++ module.xcconfig
@@ -1,12 +1,4 @@
-//
-// PLACE ANY BUILD DEFINITIONS IN THIS FILE AND THEY WILL BE 
-// PICKED UP DURING THE APP BUILD FOR YOUR MODULE
-//
-// see the following webpage for instructions on the settings
-// for this file:
-// http://developer.apple.com/mac/library/documentation/DeveloperTools/Conceptual/XcodeBuildSystem/400-Build_Configurations/build_configs.html
-//
 
-OTHER_LDFLAGS = $(inherited) -framework Accelerate -framework CoreData -framework CoreGraphics -framework CoreImage -framework CoreLocation -framework CoreTelephony -framework CoreText -framework GLKit -framework ImageIO -lc++ -lz -framework OpenGLES -framework QuartzCore -framework SystemConfiguration -framework GoogleMaps -framework GoogleMapsBase -framework GoogleMapsCore -framework GooglePlaces
+OTHER_LDFLAGS = $(inherited) -framework Accelerate -framework CoreData -framework CoreGraphics -framework CoreImage -framework CoreLocation -framework CoreTelephony -framework CoreText -framework GLKit -framework ImageIO -lc++ -lz -framework OpenGLES -framework QuartzCore -framework SystemConfiguration -framework Metal -framework GoogleMaps -framework GoogleMapsBase -framework GoogleMapsCore -framework GooglePlaces
 
 WARNING_CFLAGS = -Wno-arc-performSelector-leaks
