+++ platform/GoogleMaps.framework/Headers/GoogleMaps.h
@@ -21,16 +21,15 @@
 #import "GMSUISettings.h"
 #import "GMSCircle.h"
 #import "GMSGroundOverlay.h"
-#import "GMSMarker+Premium.h"
 #import "GMSMarker.h"
+#import "GMSMarkerAnimation.h"
 #import "GMSMarkerLayer.h"
 #import "GMSOverlay.h"
 #import "GMSOverlayLayer.h"
 #import "GMSPolygon.h"
 #import "GMSPolygonLayer.h"
 #import "GMSPolyline.h"
-#import "GMSStampStyle+Premium.h"
-#import "GMSStrokeStyle+Premium.h"
+#import "GMSStampStyle.h"
 #import "GMSStrokeStyle.h"
 #import "GMSStyleSpan.h"
 #import "GMSSyncTileLayer.h"
