+++ platform/GoogleMapsCore.framework/Modules/module.modulemap
@@ -1,6 +1,5 @@
 framework module GoogleMapsCore {
   export *
-  module * { export * }
   link "z"
   link framework "CoreData"
   link framework "CoreFoundation"
