+++ platform/GooglePlaces.framework/Headers/GMSPlace.h
@@ -9,6 +9,7 @@
 //
 
 #import <CoreLocation/CoreLocation.h>
+#import <UIKit/UIKit.h>
 
 #import "GMSPlacesDeprecationUtils.h"
 
@@ -219,6 +220,18 @@ typedef NS_ENUM(NSInteger, GMSPlacesBusinessStatus) {
  *     closed, and GMSPlaceOpenStatusUnknown if the open status is unknown.
  */
 - (GMSPlaceOpenStatus)isOpen;
+
+/**
+ * Background color of the icon according to Place type, to color the view behind the icon.
+ */
+@property(nonatomic, readonly, nullable) UIColor *iconBackgroundColor;
+
+/**
+ * The URL according to Place type, which you can use to retrieve the NSData of the Place icon.
+ * NOTES: URL link does not expire and the image size aspect ratio may be different depending on
+ * type.
+ */
+@property(nonatomic, readonly, nullable) NSURL *iconImageURL;
 
 @end
