+++ platform/GooglePlaces.framework/Modules/module.modulemap
@@ -2,16 +2,10 @@ framework module GooglePlaces {
   umbrella header "GooglePlaces.h"
   export *
   module * { export * }
-  link "c++"
   link "z"
-  link framework "CoreData"
   link framework "CoreFoundation"
-  link framework "CoreGraphics"
-  link framework "CoreLocation"
-  link framework "CoreText"
   link framework "Foundation"
   link framework "QuartzCore"
   link framework "Security"
-  link framework "SystemConfiguration"
   link framework "UIKit"
 }
