+++ titanium.xcconfig
@@ -4,7 +4,7 @@
 // OF YOUR TITANIUM SDK YOU'RE BUILDING FOR
 //
 //
-TITANIUM_SDK_VERSION = 10.0.2.GA
+TITANIUM_SDK_VERSION = 10.1.1.GA
 
 
 //
